Fix warnings in presenter

+++ friendplayer/common/Timer.cpp
@@ -4,7 +4,7 @@ void Timer::Start(long long period_us) {
     period = std::chrono::microseconds(period_us);
     auto millis = std::chrono::duration_cast<std::chrono::milliseconds>(period).count();
 
-    timer_handle = timeSetEvent(std::chrono::duration_cast<std::chrono::milliseconds>(period).count(),
+    timer_handle = timeSetEvent(static_cast<UINT>(std::chrono::duration_cast<std::chrono::milliseconds>(period).count()),
         0, &Timer::TimerCB, reinterpret_cast<DWORD_PTR>(this), TIME_PERIODIC);
     last_sync_point = std::chrono::system_clock::now();
 }
@@ -22,7 +22,7 @@ void Timer::ResetCadence() {
         timeKillEvent(timer_handle);
     }
 
-    timer_handle = timeSetEvent(std::chrono::duration_cast<std::chrono::milliseconds>(period).count(),
+    timer_handle = timeSetEvent(static_cast<UINT>(std::chrono::duration_cast<std::chrono::milliseconds>(period).count()),
         0, &Timer::TimerCB, reinterpret_cast<DWORD_PTR>(this), TIME_PERIODIC | TIME_KILL_SYNCHRONOUS);
     last_sync_point = std::chrono::system_clock::now();
 }
